requirements.txt
websocket-client==1.7.0
numpy==1.26.4
requests==2.31.0
pyzmq==23.2.0
redis==4.3.4
python-engineio==3.13.0
python-socketio==4.6.0
aws-lambda-powertools==1.25.5
websockets==13.0.1
protobuf==5.28.1